Project Update 20/02/2023

+++ src/main/java/fa/training/service/impl/CategoryServiceImpl.java
@@ -1,16 +1,14 @@
 package fa.training.service.impl;
-
-
 import fa.training.dto.CategoryDTO;
 import fa.training.entity.Category;
+import fa.training.mapper.CategoryMapper;
 import fa.training.repository.CategoryRepository;
 import fa.training.service.CategoryService;
-import fa.training.mapper.CategoryMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.NoSuchElementException;
-
+import java.util.Optional;
 @Service
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
@@ -19,7 +17,6 @@ public class CategoryServiceImpl implements CategoryService {
         this.categoryRepository = categoryRepository;
         this.categoryMapper = categoryMapper;
     }
-
     @Override
     public String addCategory(CategoryDTO categoryDTO) {
         //Convert DTO to Entity
@@ -27,7 +24,6 @@ public class CategoryServiceImpl implements CategoryService {
         categoryRepository.save(category);
         return "Add Complete";
     }
-
     @Override
     public String addCategoryList(List<CategoryDTO> categoryDTOs) {
         for (CategoryDTO categoryDTO : categoryDTOs) {
@@ -36,36 +32,31 @@ public class CategoryServiceImpl implements CategoryService {
         }
         return "Add List Complete";
     }
-
     @Override
     public List<CategoryDTO> findAll() {
         List<Category> categories= categoryRepository.findAll();
         return categoryMapper.castListEntityToDTO(categories);
     }
-
     @Override
     public Boolean deleteCategory(long categoryId) {
-       try{
+        Optional<Category> opt = categoryRepository.findById(categoryId);
+        if(opt.isPresent()) {
            categoryRepository.deleteById(categoryId);
            return true;
-       }catch (Exception e) {
+        } else {
            return false;
-       }
+        }
     }
-
     @Override
     public String editCategory(CategoryDTO categoryDTO) {
         Category category= categoryMapper.castDTOToEntity(categoryDTO);
         categoryRepository.saveAndFlush(category);
         return "Edit Complete";
     }
-
     @Override
     public CategoryDTO findByName(String name) {
         Category category = categoryRepository.findByName(name)
                 .orElseThrow(() -> new NoSuchElementException("NOT FOUND"));
         return categoryMapper.castEntityToDTO(category);
     }
-
-
 }
